+++ CHANGES.md
@@ -3,6 +3,20 @@ Changes
 
 List of changes with the latest at the top:
 
+  * v1.4.2
+    * Minor HTML tweak on options page.
+  * v1.4.1
+    * Included comment text when sending report.
+  * v1.4
+    * Styled report button.
+  * v1.3
+    * Added option to hide report button.
+  * v1.2
+    * Added 'Report Reddit Armie' button.
+  * v1.1
+    * Embedded 'Submit Reddit Armie profile' notice.
+  * v1.0
+    * Fetching list of fedora profiles from an outside source.
   * v0.9.5
     * Added footer to options page.
   * v0.9.4